Correct some glyph data offsets
Start the glyph offset at the end of the metadata table.  The table
length isn't explicitly given anywhere, so it needs to be inferred.  The
length of the metadata table is the last character value, rounded up
to the nearest multiple of 128.  Or at the very least, this is very
close to the calculation.  There might be an off-by-one error still.

Also, large characters are still broken (eg, pie chart fonts).