Major code rework
Reworked a bunch of the code to work a lot nicer and be better
organized.  There's still a bunch of cleanup to do, but this is much
closer to something that I'm OK with releasing.

Some sample fonts have been added.  The source fonts are Bert Sans and
Hack, two open source fonts.  Only a few TTF files have been included
from these two fonts.  For each of these TTF files there is a collection
of Xerox FNT font files.  Details for each file are in
sample-fonts/notes.txt.  The Xerox fonts come in two main categories:
5Word and 9700.  Both of these formats can be read by this code.